chore: clear the remaining non-dead-code clippy warnings
- deprecated egui allocate_ui_at_rect -> allocate_new_ui(UiBuilder::max_rect)
  (the old method is a literal alias for the new call)
- identical if branches merged: ACP resume retry, project rename commit,
  draft shell chip prefix (Shell chips strip $ / /sh when they run)
- type aliases for the grok bin/key/doctor caches, chip draft templates,
  the plus picker list and the realtime mint fn
- sort_by_key, clamp, strip_prefix, checked_div, digit grouping, ?-shortcuts
- tests build structs with ..Default::default() instead of field reassigns
- tests that deliberately pin design constants carry an explicit allow
- clippy.toml raises too-many-arguments to 12 for the ACP spawners
- drop the dead conn_kind helper and retarget the self-grep test that used
  it as a slice delimiter

+++ crates/grokhub-app/src/markdown.rs
@@ -142,9 +142,9 @@ mod tests {
 
     #[test]
     fn display_text_stays_inside_text_file_cap() {
-        let huge = "é".repeat((TEXT_FILE_CAP as usize) + 16);
+        let huge = "é".repeat(TEXT_FILE_CAP + 16);
         let shown = display_text(&huge);
-        assert!(shown.len() <= TEXT_FILE_CAP as usize);
+        assert!(shown.len() <= TEXT_FILE_CAP);
         assert!(shown.is_char_boundary(shown.len()));
         assert!(!shown.is_empty());
     }
